Implement Add to favorites functionaliyu

+++ frontend/src/List/EntryItem.jsx
@@ -1,22 +1,20 @@
 import React from 'react'
 import axios from 'axios';
+import { useMycontext } from '../context/MainProvider';
 import { Link } from 'react-router-dom';
 import DeleteIcon from '@mui/icons-material/Delete';
 import CheckIcon from '@mui/icons-material/Check';
+import CloseIcon from '@mui/icons-material/Close';
 import {
-    Checkbox,
+    Avatar,
+    Card,
+    CardHeader,
     IconButton,
     ListItem,
-    ListItemButton,
-    ListItemIcon,
-    ListItemText,
-    Avatar,
-    CardHeader,
-
 } from '@mui/material';
-import Card from '@mui/material/Card';
 
-const EntryItem = ({ data }) => {
+const EntryItem = ({ data, isFavorite }) => {
+    const { setFavorites, favorites } = useMycontext()
 
     const handleEdit = (id, value) => {
         axios.put(`${import.meta.env.VITE_APP_URL}/updatecheck/${id}`, { done: !value }).then(result => location.reload()
@@ -25,8 +23,13 @@ const EntryItem = ({ data }) => {
     }
 
     const handleDelete = (id) => {
-        axios.delete(`${import.meta.env.VITE_APP_URL}/delete/${id}`).then(result => location.reload()
-        ).catch(err => console.log(err));
+        if (isFavorite) {
+            setFavorites(favorites.filter((a) => a !== id));
+            // location.reload(); // use this when ready
+        } else {
+            axios.delete(`${import.meta.env.VITE_APP_URL}/delete/${id}`).then(result => location.reload()
+            ).catch(err => console.log(err));
+        }
 
     }
     return (
@@ -37,13 +40,13 @@ const EntryItem = ({ data }) => {
                         <Avatar
                             onClick={() => handleEdit(data._id, data.done)}
                             sx={{ bgcolor: data.done ? 'green' : '' }}
-                            aria-label="recipe">
+                            aria-label="entry">
                             {data.done ? (<CheckIcon />) : <div />}
                         </Avatar>
                     }
                     action={
                         <IconButton aria-label="settings" onClick={() => handleDelete(data._id)}>
-                            <DeleteIcon />
+                            {isFavorite ? (<CloseIcon />) : (<DeleteIcon />)}
                         </IconButton>
                     }
                     title={<Link to={`/entry/${data._id}`}>{data.name}</Link>}
